Rename report components
